requirements-dev.txt
-r requirements.txt
pytest==8.0.0
django-debug-toolbar==4.3.0
tqdm
